backend/requirements.txt
flask 
python-dotenv 
langchain 
langchain_groq
pyyaml